Very basic routing working
Resolved #4. This has a very crude implementation of the routing desired. However, it's not working optimally. For one thing, each change in URL seems to cause a refresh, which isn't desired or expected. This actually is probably a side effect of using a direct browser address bar manipulation. Also, discovered a bug in the calendar logic that will not allow Januaries to render (opened issue #7). But the bug isn't impeding the operation of the routing itself.