Added methods on `CalendarService` to calculate next and previous months
This will allow the component to ask the service for the values of the next and previous months, which it can then make available as part of the navigation controls. Also did a little refactoring of the service's internal state into a separate value holder object. Objects of this type are then returned from the calculation methods.

ref issue #5

(cherry picked from commit dce594a)