Completely revamped the `CalendarService`
As part of implementing issue #5, I realized that I needed to deal with the mismatch between "friendly" months (1-based), and "real" months (0-based, which is used by various JS calendar implementations). In order to achieve that, I totally refactored the `CalendarService` and surrounding module to distinguish between the real values, which are used internally, and friendly values, which are used in client-facing contexts.

Along with the refactoring, I provided documentation comments. I also took the opportunity to alter the API to be less verbose, easier to use, and more idiomatic. This resulted in changes to many unit tests and components. However, the end result is that all existing functionality remains in tact, and I feel that the service is easier to use and more reusable. It also allows me to finish implementing #5 without propagating the mismatch throughout the code.