Changed the `MonthNavigationComponent` to route directly to the desired month
Rather than emitting an event that some other component listens to, the `MonthNavigationComponent` now has a reference to the `Router` and directly navigates to the month that it is configured with. The routing is working properly (URL bar is updated), and the navigation controls update. The header and day grid do not update yet.

ref issue #5