Mapped store getters to computed component properties
Created computed properties for the `Header` and `CalendarGrid`
components that are mapped to the Vuex getters. This restored the
calendar functionality for displaying the current month. The month can
be selected via routing parameters, including the "default" route that
redirects to the current month.

Added a stub for the `beforeRouteUpdate` handler on the `Calendar` view,
and fixed a compiler warning.