Added support for setting locale from query params
This exposed an issue with getting weekday names for locales that do not
use Sunday as the first day of the week. The rotation routine was
modifying Moment's underlying weekday array, so subsequent calls were
continuing to rotate it. Fixed by taking a copy of the array before
performing the rotation. Added a new unit test for this behavior.

Also did a little cleanup and added some code reminders.