Reverted some dependency upgrades and fixed a unit test
Found that upgrading dependencies caused compilation errors, possibly
due to some incomplete downstream upgrades to the Babel 7.1 release.

Reverted the change on the Vuex store that switched a default export to
a named export. Found that the issue with ESLint was caused by the same
dependency upgrades. This also fixes the failing unit test.