Complete state refactoring
Backed off of using a Moment object to hold the selected month, because
the locale stuff got very complicated. Instead, defined an actual class
to hold the year/month combo and moved some date manipulation logic to
it. Extracted all of the locale handling to the getters, which will
combine the locale options with the date manipulation, when needed.

Also expanded the documentation on all functions and classes.